requirements.txt
Flask
Werkzeug
python-dotenv

setuptools
pillow
qrcode[pil]
